add 2 APIs (dropdown) that's will help in distribution functionalities

+++ src/main/java/com/eden/eden_crm_sec_crm_back/dto/lookup/LKCustomerContractServiceDto.java
@@ -2,7 +2,6 @@ package com.eden.eden_crm_sec_crm_back.dto.lookup;
 
 
 import com.eden.eden_crm_sec_crm_back.base.dto.BaseDto;
-import com.eden.eden_crm_sec_crm_back.dto.CustomerServiceDetailsDTO;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -18,11 +17,11 @@ import java.io.Serializable;
 @JsonIgnoreProperties
 public class LKCustomerContractServiceDto extends BaseDto<Long> implements Serializable {
     private Long quantity;
-    private Long unitPrice;
+    private Double unitPrice;
     private Long serviceDetailsId;
     private Long customerContractId;
 
-    public LKCustomerContractServiceDto(Long id, Long quantity, Long unitPrice,
+    public LKCustomerContractServiceDto(Long id, Long quantity, Double unitPrice,
                                         Long serviceDetailsId, Long customerContractId) {
         this.setId(id);
         this.quantity = quantity;
